+++ Escape/src/model/backbone/building/elements/Sign.java
@@ -17,7 +17,7 @@ public class Sign {
 		end = b;
 	}
 	
-	public Sign(double x1, double y1, double x2, double y2){
+	public Sign(int x1, int y1, int x2, int y2){
 		begin = new Point(x1, y1);
 		end = new Point(x2, y2);
 	}
